third_party/blink/web_tests/fast/dom/shadow/shadow-tree-styles-select-host-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test for issue 247275: shadow tree styles selects shadow host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test that tag selectors in a shadow tree cannot match without :host.
PASS borderColorOf('host') is "rgb(0, 0, 0)"
PASS borderColorOf('host-child') is "rgb(0, 0, 0)"
Test that * in a shadow tree cannot match without :host.
PASS borderColorOf('host') is not "rgb(0, 128, 0)"
PASS borderColorOf('host-child') is "rgb(0, 0, 0)"
Test that :not(*) in a shadow tree cannot match without :host.
PASS borderColorOf('host') is "rgb(0, 0, 0)"
PASS borderColorOf('host-child') is "rgb(0, 0, 0)"
Test that :first-child in a shadow tree cannot match without :host.
PASS borderColorOf('host') is "rgb(0, 0, 0)"
Test that styles in a containing treescope wins if specificities are the same.
PASS borderColorOf('host') is "rgb(0, 128, 0)"
Test that rules which has higher specificity win.
PASS borderColorOf('host') is "rgb(255, 0, 0)"
Test that child selector with a shadow host and children of a shadow root of the shadow host works.
PASS borderColorOf('divA') is "rgb(0, 0, 0)"
PASS borderColorOf('divA-child') is "rgb(0, 0, 0)"
PASS borderColorOf('host') is "rgb(0, 0, 0)"
PASS borderColorOf('host/divC') is "rgb(0, 128, 0)"
PASS borderColorOf('host-child') is "rgb(0, 0, 0)"
Test that descendant selector with a shadow host and descendants of a shadow root of the shadow host works.
PASS borderColorOf('host/spanD') is "rgb(255, 0, 0)"
Test the case where there are no styles in a contanining treescope.
PASS borderColorOf('host/target') is "rgb(0, 128, 0)"
Test that style sharing is disabled when styles in shadow trees would match shadow hosts.
PASS borderColorOf('hostA') is "rgb(255, 0, 0)"
PASS borderColorOf('hostB') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
